chore(org): Don't load mdx dynamically in a hook
